refactor: move fluid_stream and refactor tests
