+++ src/ecalc_neqsim_wrapper/thermo.py
@@ -20,8 +20,8 @@ from ecalc_neqsim_wrapper.mappings import (
 )
 from libecalc.common.decorators.capturer import Capturer
 from libecalc.common.logger import logger
-from libecalc.domain.process.entities.fluid_stream.eos_model import EoSModel
-from libecalc.domain.process.entities.fluid_stream.fluid_composition import FluidComposition
+from libecalc.domain.process.value_objects.fluid_stream.eos_model import EoSModel
+from libecalc.domain.process.value_objects.fluid_stream.fluid_composition import FluidComposition
 
 STANDARD_TEMPERATURE_KELVIN = 288.15
 STANDARD_PRESSURE_BARA = 1.01325
