+++ src/libecalc/domain/process/compressor/core/results.py
@@ -11,8 +11,8 @@ from libecalc.domain.process.core.results.compressor import (
     CompressorTrainCommonShaftFailureStatus,
     TargetPressureStatus,
 )
-from libecalc.domain.process.entities.fluid_stream.fluid_composition import FluidComposition
 from libecalc.domain.process.value_objects.chart.chart_area_flag import ChartAreaFlag
+from libecalc.domain.process.value_objects.fluid_stream.fluid_composition import FluidComposition
 
 
 class CompressorTrainStageResultSingleTimeStep:
