+++ src/libecalc/domain/process/value_objects/fluid_stream/__init__.py
@@ -1,0 +1,15 @@
+from libecalc.domain.process.value_objects.fluid_stream.conditions import ProcessConditions
+from libecalc.domain.process.value_objects.fluid_stream.eos_model import EoSModel
+from libecalc.domain.process.value_objects.fluid_stream.fluid_composition import FluidComposition
+from libecalc.domain.process.value_objects.fluid_stream.fluid_stream import FluidStream
+from libecalc.domain.process.value_objects.fluid_stream.mixing import SimplifiedStreamMixing
+from libecalc.domain.process.value_objects.fluid_stream.thermo_system import ThermoSystemInterface
+
+__all__ = [
+    "FluidStream",
+    "ProcessConditions",
+    "SimplifiedStreamMixing",
+    "ThermoSystemInterface",
+    "EoSModel",
+    "FluidComposition",
+]
