+++ src/libecalc/domain/process/value_objects/fluid_stream/eos_model.py
