+++ src/libecalc/domain/process/value_objects/fluid_stream/exceptions.py
