+++ tests/libecalc/domain/process/fluid_stream/test_conditions.py
@@ -1,6 +1,6 @@
 import pytest
 
-from libecalc.domain.process.entities.fluid_stream.conditions import ProcessConditions
+from libecalc.domain.process.value_objects.fluid_stream.conditions import ProcessConditions
 
 
 class TestProcessConditions:
