+++ tests/libecalc/dto/test_energy_model.py
@@ -12,10 +12,10 @@ from libecalc.domain.component_validation_error import (
     ProcessChartValueValidationException,
     ProcessEqualLengthValidationException,
 )
-from libecalc.domain.process.compressor import dto
 from libecalc.domain.infrastructure.energy_components.turbine import Turbine
-from libecalc.domain.process.entities.fluid_stream.fluid_composition import FluidComposition
+from libecalc.domain.process.compressor import dto
 from libecalc.domain.process.value_objects.chart.generic import GenericChartFromDesignPoint, GenericChartFromInput
+from libecalc.domain.process.value_objects.fluid_stream.fluid_composition import FluidComposition
 from libecalc.presentation.yaml.model import YamlModel
 from libecalc.presentation.yaml.model_validation_exception import ModelValidationException
 from libecalc.testing.yaml_builder import YamlTurbineBuilder
